LayoutTests/platform/chromium/media/media-can-play-flac-audio-expected.txt

Test HTMLMediaElement canPlayType() method with multiple .flac MIME types.

These tests may be expected to fail if the WebKit port does not support the format.

EXPECTED (audio.canPlayType('audio/x-flac') == 'maybe'), OBSERVED '' FAIL
EXPECTED (audio.canPlayType('audio/flac') == 'maybe'), OBSERVED '' FAIL
END OF TEST
